+++ changelog.md
@@ -4,6 +4,9 @@ Codion Change Log
 ## 0.18.10
 ### is.codion.common.core
 - DatabaseException.statement() and sqlState() now return Optional.
+### is.codion.swing.common.ui
+- FilterTable.filterPanel() now returns TableConditionPanel.
+- FilterTable.Builder.filterPanelFactory added.
 ### is.codion.framework.domain.db
 - Module is.codion.framework.domain.db split from is.codion.tools.generator.domain.
 - DatabaseDomain renamed SchemaDomain.